another try to fix the event form
