+++ components/KidProfile/index.js
@@ -2,36 +2,24 @@ import Image from "next/image";
 import { StyledKidCard, StyledEditButton, StyledBackLink, StyledEventList } from "./KidProfile.styled";
 import { formatDate } from "../../resources/dateUtils";
 import EventForm from "../EventForm";
-import useSWR from "swr";
-import { useRouter } from "next/router";
 
-export default function KidProfile({ handleSubmit }) {
-  const router = useRouter();
-  const { id } = router.query;
+export default function KidProfile({
+  kidData, 
+  handleSubmit }) 
+  {
 
-  const { data, isLoading, mutate} = useSWR(`/api/kids/${id}`);
-  
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (!data) {
-    return <div>Where are the kids???</div>;
-  }
-
- {
   return (
       <StyledKidCard>
         <Image
           width={200}
           height={200}
-          src={data.image}
+          src={kidData.image}
           alt="Kid Photo"
         />
-        <h2>{data.name} </h2>
-        <h3>* {formatDate(data.birthDate)}</h3>  
+        <h2>{kidData.name} </h2>
+        <h3>* {formatDate(kidData.birthDate)}</h3>  
       <StyledEventList>
-      {data?.events?.map((event) => (
+      {kidData?.events?.map((event) => (
           <li key={event._id}>
            <p>{event.title}</p>
            <p>{formatDate(event.date)}</p>
@@ -42,4 +30,4 @@ export default function KidProfile({ handleSubmit }) {
       <StyledBackLink href={"/"}>Zurück</StyledBackLink>
       </StyledKidCard>
   );
-}}
+}
